requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-llms-gemini
python-dotenv